pyproject.toml
[project]
name = "eedom"
version = "0.2.0"
description = "Eagle Eyed Dom — fully deterministic dependency and code review for CI"
requires-python = ">=3.12"
dependencies = [
    "click==8.3.3",
    "pydantic==2.13.3",
    "pydantic-settings==2.14.0",
    "structlog==25.5.0",
    "httpx==0.28.1",
    "psycopg[binary]==3.3.3",
    "psycopg-pool==3.3.0",
    "orjson==3.11.8",
    "packaging==26.1",
    "pyarrow==24.0.0",
    "agent-framework-github-copilot==1.0.0b260423",
    "jinja2==3.1.6",
    "pyyaml==6.0.3",
    "watchdog==6.0.0",
    "rich>=13.7.0",
    "starlette>=1.0.0",
    "uvicorn>=0.46.0",
]

[project.scripts]
eedom = "eedom.cli.main:cli"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest==9.0.3",
    "pytest-asyncio==1.3.0",
    "black==26.3.1",
    "ruff==0.15.11",
    "respx==0.23.1",
    "hypothesis==6.152.1",
    "packaging==26.1",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.black]
target-version = ["py312"]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
